- separated syscalls and hi-level functions with allocation logic - added calling PdhGetCounterInfoW - added calling PdhExtractWildCardPathW - added missing functions comments
